pyproject.toml
[tool.poetry]
name = "stockbot"
version = "0.1.0"
description = ""
authors = ["Acusick1 <andrewcusick13@outlook.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
requests = "^2.27.1"
h5py = "^3.6.0"
yfinance = "^0.1.70"
pandas = "^2.2.2"
numpy = "^1.26.4"
matplotlib = "^3.5.2"
tables = "^3.7.0"
streamlit = "^1.9.2"
seaborn = "^0.11.2"
pydantic = "^2.7.0"
pandas-market-calendars = "^4.0"
Backtesting = "^0.3.3"
hyperopt = "^0.2.7"
pandas-datareader = "^0.10.0"
wrapt = "^1.15.0"
ta = "^0.10.2"
scikit-learn = "^1.2.2"
yellowbrick = "^1.5"
psycopg2 = "^2.9.6"
sqlalchemy-timescaledb = "^0.4"
sqlalchemy-utils = "^0.41.0"
pandarallel = "^1.6.4"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-mock = "^3.9.0"
ipykernel = "^6.21.3"
ruff = "^0.3.7"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
